Major server refactoring. awsController is now split into ec2Instances and cloud controllers. Types and interfaces are up to date. General clean-up and error fixes.
